Split two demos into two folders:  Internet Data Display and Image Slide Viewer
