--- circuitPython/examples/Internet_display_and_ir_sensors/Readme.txt
@@ -1,6 +1,0 @@
-This release include below demo:
-    1. Demo 1rt: Internet data display Demo
-       See apps/internet_display/internet_display_readme.txt
-    
-    2. Demo 3rd: 3.	Image Slide show demo
-       See apps/ir_sensors_and_gestures_readme.txt
