+++ circuitPython/examples/image_slide_viewer/lib/adafruit/adafruit_irremote.py
