+++ circuitPython/examples/image_slide_viewer/test/DemoImageSlideShow/jpg/converter.sh
