+++ circuitPython/examples/internet_data_display/lib/adafruit/adafruit_esp32spi/adafruit_esp32spi_wsgiserver.py
