+++ circuitPython/examples/internet_data_display/lib/adafruit/adafruit_io/adafruit_io_errors.py
