+++ circuitPython/examples/internet_data_display/page2_ui.py
