+++ circuitPython/examples/internet_data_display/secrets.py
@@ -1,0 +1,9 @@
+wifi_secrets=[
+    {'ssid':'Tuan', 'password':'0973907059'},
+    {'ssid':'nah', 'password':'abc12345'},
+]
+
+aio_secrets={
+    'aio_username':'tuan_nguyen_brtchip',
+    'aio_key':'aio_JPVI82dpB'+''+'BNy4MiqDUn7YEZeCIky'
+}
